requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap-v5==1.0.7
django-environ==0.8.1
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-leaflet==0.27.0.dev0
django-registration==2.4.1
django-tinymce==3.4.0
django-utils-six==2.0
gunicorn==20.1.0
importlib-metadata==3.10.1
Pillow==8.4.0
psycopg2==2.9.2
psycopg2-binary==2.8.3
python-decouple==3.5
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
